data-warehouse/pyproject.toml
[project]
name = "data-warehouse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-storage>=3.2.0",
    "pandas>=2.3.1",
    "pyarrow>=21.0.0",
]
